teleport: organize importer api
