+++ packages/teleport/src/api.ts
@@ -1,4 +1,5 @@
-export * as importer from './importer/api.js';
+// `public` refers to the importer type; this file also exports the raw importer to avoid an additional barrel file.
+export * as importer from './importer/public.js';
 export * as mutex from './mutex.js';
 export * as exporter from './exporter.js';
 export * as interfaces from './shared/interfaces.js';
